Added lexer for new Hy language, a variant of Lisp running on Python.  See: http://hylang.org
Note that this shares a file extension with the Hybris language
but the analyse_text() function does a good job of making the
distinction.

This did however require one change to the tests to actually
pass the code of the test file to get_lexer_for_filename()
so that the tests would differentiate Hy from Hybris.

And while this is a Lisp and shares some syntax with Clojure,
it has been added to the agile.py file to share the lists of
keywords and builtins with the PythonLexer.