Varnihs Lexer: Add various missing tokens
The lexing is a bit imprecise right now, but other than that I think the
biggest missing parts right now is the inner workings of ACLs and vmods.

It might be missing a few variables too.

In general, the lexer is in need of some reorganizing to tidy things up.