Add support for the Fennel programming language
This is a pretty straightforward language in the lisp family with a
small number of special forms. Since Fennel runs on the Lua runtime,
the list of builtins is the same as that of Lua, so it might be
possible to re-use the definition from the Lua lexer, but since I
don't know Python I couldn't figure out how that would work; maybe
someone else could add that.