Create a Lexer class for BBC Basic files.
The Lexer class for BBC Basic handles both the numbered lines, and
unnumbered lines, of the detokenised (text) format of BBC BASIC.
The tokeniser copes, in a naive manner, with the orignal versions,
and BASIC V. It does not handle other extensions at this time, nor
does it handle inline assembler. This should be sufficient for most
cases where code needs to be presented in a colourful manner.