sql lexer: Add temporal support keywords
SQL:2011 introduced support for temporal databases[1] via new keywords to
create and manage temporal tables. This patchset adds these new keywords to
the sql lexer.

One concrete example of these feature is the temporal tables support in
MariaDB[2]. As evident from the code listing on that page, the syntax
highlighting for these keywords is not as widespread as expected (I wouldn't
be surprised if code listing on that page ultimately use pygments for syntax
highlighting !)

[1] https://en.wikipedia.org/wiki/SQL:2011
[2] https://mariadb.com/kb/en/temporal-data-tables/