Update SystemVerilog literal constants
The original implementation was missing some of the more arcane features
such as underbars, the character 's' for signed/unsigned, support for
spaces before/after the base specifier, capital letter base specifiers
(ie 'B 'D 'H), and the allowed, 4-state 'xXzZ?' characters.

For regular integers, the 'l' and 'L' suffixes are not valid,
and they also support underbars.

Create a new test that exercises most of the interesting kinds of
SystemVerilog numbers.