Matlabsession line continuation (#1399)
* Add explicit line continuation for Matlab session

Matlab lines can be explicitly continued with the ... syntax at the end
of a line. In the Session lexer, this requires continuing to the next
line to add more text. Otherwise, the next line is marked as output.

* The ellipses in Matlab should be a Keyword

The built-in Matlab syntax highlighter highlights ... with the same
formatting as a keyword. Everything after that on the line should be a
comment.

* Update Matlab functions and keywords from R2018a

* Fix many spaces in assignment formatted as string

In command mode, MATLAB allows mutiple space separated arguments to a
function which are interpreted as char arrays, and are formatted as
Strings. This check was also catching cases where there were multiple
spaces following an assignment or comparison operation and formatting
the rest of the line as a string. Now, if an = or operator is found, the
commandargs state is popped and control returns to the root state.

* Add tests for MATLAB formatting