SystemVerilog keyword/operator improvements (#1464)
* Move SystemVerilog type keywords

Put them next to the generic keywords list.

* Change a couple SystemVerilog keywords to operators

The 'inside' and 'dist' keywords are described as operators in the
SystemVerilog standard, below unary increment/decrement, and above
concatenation in precedence.
See 1800-2017 tables 11-1 and 11-2 for a list of operators.

This matches the description of pygemnts Operator.Word token:
"For any operator that is a word (e.g. not)."

* Add a SystemVerilog operators unit test

Copy/paste the contents of 1800-2017 Table 11-2,
and see what the SV lexer chops it up into.

I made lots of comments for potential improvements.
Some operators, such as '[' and '.' are being labeled as punctuation.
Also, multi-character operators such as '<<<=' are being split up
into multiple, single-character tokens, eg '<' '<' '<' '='.