Improve SystemVerilog operators/punctuation
The most important fix is to update the lexer so it recognizes tokens
like '<<<=' instead of '<' '<' '<' '='.

There are two minor changes to functionality:
1. Change '.' from Punctuation to Operator, and
2. Add '$' to the list of Punctuation

Python treats '.' as an operator. SystemVerilog uses the '.' for
basically the same things, such as class attributes, class or built-in
methods, eg 'array.push_front', and struct/union members.
Given the infix notation, this seems like the correct convention.

The '$' character is used by the SVA (assertion) mini-language.
Mark it as punctuation for lack of a better category.

Note that the ':' character is an operator b/c it is part of the ternary
operator. However, it also used as punctuation in other places, eg
'endmodule : identifier'. I left a note to re-address this.