Update SystemVerilog module/endmodule groups
Following the same pattern as class/endclass, we have:
module identifier;
   ...
endmodule [: identifier]

Update the module/endmodule keywords to the slightly more specific
Keyword.Declaration, change the ':' character to Punctuation,
and use the generic 'Name' for the module's name.
Neither Name.Class or Name.Namespace seem to be correct here,
so just go with the more generic choice.