LaTeX: \catcode assignment should be terminated
This is almost irrelevant. But \catcode`\_=8 without termination might
let LaTeX expand prematurely its token stream. Of course here in the
context of fancyvrb.sty chances are 99.99% that what comes next does not
expand or expands safely. I did not look in code.

Only a matter of style, to be on the safe side.