LaTeX: avoid uneven baselines in case highlighting uses \fcolorbox (#1708)
Without this, the \strut causes the background color to apply to a full
baseline height, then the framing done by \fcolorbox augments beyond
that the box height and as a result the line where this happens will
have extra distance from previous and next lines.

Correct that by reducing the apparent box size by exactly the width of
the frame (\fboxrule).

Also, make the change to \fboxsep local. Reason: if texcomments is True,
some arbitrary LaTeX mark-up can be executed in a later part and the
setting of \fboxsep could modify output; macro \<cmdprefix>@bc is
executed at top level, cf. \<cmdprefix>@do.