Fix rare `missing number treated as zero`
I have a `.tex` document that uses a lot of packages. One of them (or a combination thereof) must mess with the `-` symbol `\setlength` (as defined in the calc package), resulting in:
`! Missing number, treated as zero`
on `\PYG{err}{XXX}` with the default highlighting.
Replacing `\setlength{\fboxsep}{-\fboxrule}` with the seemingly more robust `\setlength{\fboxsep}{0pt minus \fboxrule}` fixes the issue.

I do not have a MWE exhibiting the bug.