Update LLVM lexer to use keywords of LLVM 12.0. (#1824)
The keywords are based on the extraction of keywords the LLVM 12.0 lexer
source code. They were taken from the `KEYWORD` and `INSTKEYWORD` macros
for the 'Keyword' tokens, and 'Type' tokens were from the `TYPEKEYWORD`
macro. The keywords were then sorted against the current set.

Due to formatting issues in the arrays, the actual Type diff is:

``` diff
+x86_mmx
+x86_amx
```

and the Keyword diff is:

``` diff
+aarch64_sve_vector_pcs
+aarch64_vector_pcs
+alwaysInline
+amdgpu_gfx
+blockcount
+byref
+callbr
+canAutoHide
+cfguard_checkcc
+dso_local_equivalent
+fneg
+mustprogress
+nocallback
+nofree
+noInline
+nomerge
+noprofile
+nosync
+noundef
+null_pointer_is_valid
+param
+params
+partition
+preallocated
+sanitize_memtag
+speculative_load_hardening
+tailcc
+typeidCompatibleVTable
+varFlags
+vcall_visibility
+virtFunc
+vscale
+vTableFuncs
+x
```