Add lexer and style for LilyPond
GNU LilyPond is a text-based music typesetter. Because its concepts
are completely different from programming languages, this adds
a special-purpose to highlight the special tokens.

The SchemeLexer is expanded to gain the ability of detecting when
one expression stops. LilyPondLexer subclasses SchemeLexer.

Builtins (the most important part) are generated with a script
put in external/ (as was already done for Lasso).

As part of this change, the CPSALexer is made no longer to
inherit from SchemeLexer. The inheritance was unused anyway.