NF: adding an example of use of simple filter
@simplefilter is great, but also not very intuitive. Indeeds, the syntax seems
to indicate that you define a function with four arguments while in reality you
define a class whose constructor takes arbitrary keyword arguments. I believe in
this case an example to show how to instantiate this filter is really necessary.

Regarding simplefilter, I also believe that it could be improved in two simple
ways:
* accepting any method which takes lexer and stream as a filter. That would be
  sufficient as long as there is no option
* the @simplefilter decorator could deal with `self` so that the user do not
  have to add it themselves. Probably not worth doing it no, as it would break
  compatibility with current version, but would be even simpler to use