Make test_contrasts.py ignore token.Whitespace (#1938)
With the VisibleWhitespaceFilter Pygments can be configured to replace
whitespace characters like spaces, tabs and newlines with Unicode
characters like ·, » and ¶ respectively.

The min_contrast metric is meant to represent the readability of a
style. Increasing it should mean that the style becomes more readable.

This commit therefore excludes the token.Whitespace contrast from the
min_contrast calculation, since giving whitespace characters a high
contrast would make them so intrusive that they would actually hinder
the readability of the source code.