Add missing colors to dark styles
Some dark styles did not define a color for every token type,
resulting in black text (the browser default for text) on dark
backgrounds (defined by the styles) unless the web page had some
CSS to remedy that like:

  body { color: white; background: black; }

We however don't want the readability of styles to rely on external CSS.

Part of #1718. Fixes some unreadable styles reported in #1526.