Fix highlighting for alloy facts
Alloy (http://alloytools.org/) facts can take a string token in addition
to an identifier. This is because the name isn't actually used for
facts, it's just for readability.

This is technically undocumented behavior, but it's widely used in
legacy specs and is handled this way in the official Alloy IDE.