Fix handling of " as a token
Legacy alloy uses `'` to distinguish variables:

    all b, b': Foo | bar

But since Alloy 6 makes `'` into an operator, the new recommendation is
to use `"`:

    all b, b": Foo | bar

See
https://alloytools.org/alloy6.html#compatibility-with-pre-6-models.

The lexer currently reads `b"` as starting a string. This commit fixes
this by only having `"` begin a string if it's the start of a token.

Additionally, `'` is removed from identifiers.