demo: make form usable while Pyodide is loading
Previously the form could only be filled out after Pyodide finished
loading (since the <select> options were generated via Pyodide).

This commit changes this, so that the <select> options are already
in the static HTML. The "Loading Python..." form overlay is removed,
instead just the Highlight button is disabled, letting users fill
out the form while Pyodide is loading.

When a ?code query parameter is given a "Loading Python..." message
is displayed below the form, since in that case users probably want
to wait for the highlighted code (instead of filling out the form).