Java: fix lexing of 'record' soft keyword
Refactor the Java lexer to treat `record` as a soft keyword.

Previously, the lexer assumed record is a reserved word, even though
it is a soft keyword which can be used as a variable name.

This refactor lexes record as a keyword only if it appears at the
beggining of the line, with some potential other keywords like public
and private preceding it.