Add Make target for building Pyodide with currently checked out Pygments
Add a target to the documentation's Makefile to make Pyodide with
Pygments from the current checkout.

Additionally, use this target when building the documentation with
the demo page.

The new target works by building Pyodide on a Docker container, and
then exporting the build artifacts back to the host.

The Dockerfile used by the new target is based on the Dockerfile which
was used for building Pyodide on the CI:
https://github.com/pygments/pyodide-artifacts/blob/master/container/Dockerfile

The main difference between the two is that the new Dockerfile uses the
newest Pyodide base image, which is built by a repository which is actively
maintained (the Iodide project is no longer maintained).

The purpose of this change is to allow running the demo locally,
which was not possible previously, as Pyodide was only built by the
CI when deploying to Github Pages.