Merge mapping file generation scripts (#2152)
Use a unified script, to reduce code duplication and in preparation
for doing a similar thing with styles and filters.  The new script
also uses a bit more modern Python APIs (e.g., pathlib).

Unlike the previous scripts, it does not replace replace CRLF with LF
because Git should do that itself.