Detect true-color terminal support for pygmentize (#2160)
For the pygmentize command line, this pull request adds detection of true-color support on the terminal and makes that the default formatter. It looks up the COLORTERM environment variable having a value of truecolor or 24bit. If that is the case, it sets the default formatter to use the TerminalTrueColorFormatter.

The COLORTERM environment variable is a convention that has been adopted by some terminals for specifying true-color is supported.