Markdown: add comment parsing to lexer
Prior to this commit, comments were not included in the Markdown lexer.
This patch adds comments to the Markdown lexer output. Now comments such
as`<!-- comment -->` will be highlighted as comments appropriately. This
includes multi-line comments.

The regex strings for Markdown headings have been altered to allow for
comments at the end of heading lines.