ShellSessionBaseLexer: Support whitespace in virtualenv prompt
The default virtualenv prompt contains a single space after
the name of the virtualenv in parentheses.

Commit 3f403687 which resolved #1253 assigns the "Text" class to
any whitespace between the virtualenv name and the rest of the prompt,
but this is problematic, because it makes the space selectable, and
doesn't include it in the prompt.

Fix this by assigning the Generic.Prompt class to the space between the
name of the virtualenv and the rest of the prompt, instead of Text.
Also update the test accordingly.

Closes #2496

Signed-off-by: Vangelis Koukis <evangelos.koukis@hpe.com>