ShellSessionBaseLexer: Only match prompt once
By default, ShellSessionBaseLexer attempts to detect multiple prompts
within the input. This works nicely when input contains multiple
commands, but has problems when the output of the commands contain
things that looks like prompts, e.g., commented-out lines that
start with "#".

There already exists a test for this [prompt_in_output.txt],
but it contains the secondary prompt, not the primary prompt. Add a test
which contains the primary prompt as well [prompt_in_output2.txt].

Since there is no guaranteed way to actually know which lines are part
of the output and which lines actually correspond to prompts,
add an option to support only matching the prompt once, and stop parsing
for prompt when we've already matched a prompt once.
This allows us to optionally standardize on a single command per console
block, and remove the ambiguity.

Closes #2497

Signed-off-by: Vangelis Koukis <evangelos.koukis@hpe.com>