A couple small fixes in the OCaml lexer. (#2521)
* Remove "value" from the keywords.

"value" is not a keyword in OCaml.
It was once a keyword in the so-called "revised" syntax of OCaml,
but this syntax was never mainstream and is now essentially extinct.

* Add "and" to the keywords.

"and" is a keyword in OCaml.
It is used e.g. in "let x = 0 and y = 1".
It is not an infix operator.