Add files via upload
Solving issue 2293:
When encountering the multi-line comment start marker %{, use #push to push the current state onto the stack and enter a new comment state. This allows handling nested multi-line comments. Use #pop to pop the current state from the stack and return to the previous state. This ensures that nested comments are correctly closed and the lexer returns to the appropriate state.