DebianControlLexer: Fix issue with trailing commas on depends lists
On the deb822 format[0] that d/control and many other Debian-specific
files use, a multiline field is continued by a whitespace at the end of
the line. This commit fixes an issue where a trailing comma at the end
of a dependency list made Pygments believe the following field was a
part of the dependency list. For example, with the following code:

  Depends:
   foo,
   bar,
   baz,
  Other-Field: other-value

The Other-Field field was highlighted to be a part of the Depends field
(which it was not). This is fixed with this commit. A similar issue
still exists with the with the Uploaders field.

[0]: https://manpages.debian.org/unstable/dpkg-dev/deb822.5.en.html

Signed-off-by: Taavi Väänänen <taavi@debian.org>