Add a TableGen lexer
This commit adds a basic lexer for LLVM's TableGen DSL[1][2].

The major design decision is about how to tokenize "code blocks" since
those are, at a TableGen language level, string literals. In practice,
however, in LLVM 99% of the time they are snippets of C++ which are
processed by a TableGen backend. This would ideally be a configurable
property of the lexer. For the initial implementation, it is assumed
that code blocks contain C++ code and so the CppLexer is invoked on
their contents.

[1] https://llvm.org/docs/TableGen/index.html
[2] https://llvm.org/docs/TableGen/ProgRef.html