fix: switch to suggested '--strip-escapes=?' option.
Adds support for '--strip-escape={off,on,auto,raw,semi-raw}'.
The first 3 are as according to the suggestions from the comments
in PR #2492.

The test case throws up errors for the 'off' and 'semi-raw' values
as output cuts off after 32808 characters.  I'm fairly certain the
bug is in the test, as running the program on the same data with the
same options outside of `pytest` works fine.