fix: adds flush after last write in `main_inner`
Also, modifies `tests.test_cmdline.test_strip_escapes_opt` to use a
shorter file in testing.  Not only is a longer file not required to
demonstrate the issue for which the test was added, the larger values
obscure what is happening when the test fails.

The test now passes for all cases.