test: adds a test for issue #2692
This issue involves the xml lexer mistagging a DOCTYPE containing
an inline DTD.  The closing part of the DOCTYPE is recognised as
`Text`.

The test added calls `pytest.xfail` if it detects the situation
described when processing the sample of code provided:

```xml
<?xml version="1.0" encoding="UTF-8"?>
<!DOCTYPE hello [<!ELEMENT hello (#PCDATA)>]>
<hello>Hello Word!</hello>
```

The test is currently slightly fragile, in that it will raise
an exception if "]>" is not part of a single token; this is
easily fixed, but does not cause a problem with the current lexer,
and is unlikely to do so in the future.