fix: replaces pytest for issue #2806 with snippet
This change replaces the test written using pytest with the file
`tests/snippets/csharp/test_file_keyword.txt`.

The change removes the ability to toggle between `xfail` and real
failure for cases that are not expected to pass, but brings testing
in line with the rest of the codebase.  The tests involving `await`
have been removed - which is fitting, as it was not the same issue.