fix+feat: adds support for inline dtds to XmlLexer
This is a WIP that enables correct processing of DOCTYPE declarations
with inline DTDs.  It modifies `pygments.lexers.html.XmlLexer` so that
it processes the DTD sections, assigning `Token.Other` to everything
within the '[...]' (including the brackets).

Changing the output so that `Other` is replaced by `Comment.Preproc`
fixes #2692.

The commit also adds a class `XmlWithInlineDtdLexer`, derived from
`DelegatingLexer`, which causes the material within the inline DTD
to be processed by `DtdLexer`.

There are, as of yet, no tests.  Nor is the lexer able to be found
when running `python -m pygments -lxml+dtd -fraw` (or any variant).
I have not looked into this issue.