fix: adds support for xml + xml+dtd lexers
Still very much a WIP, this renames the revised `XmlLexer` class to
`XmlLexerBase` and makes that a base class of an all-new `XmlLexer`.
`XmlLexer` now generates output similar to the verion on master,
replacing `Other` tokens with `Comment.Preproc`.  The difference in
output is that this one splits the DOCTYPE into several tokens and fixes
issue #2692.  `XmlWithInlineDtdLexer` passes off the inline DTDs to
`DtdLexer` for processing, producing interesting results.  Tests are
added, a previously added test removed, and "tox -e mapfiles" was run
so that the new lexer can be invoked by name.