Remove thousands separators from line numbers
Line numbers are rarely seen with thousands separators — a statement's
location is referred to as "line 1304" ("line thirteen-oh-four"),
much more often than "line 1,304" ("line one thousand, three hundred
four").

More importantly, multiple line numbers are displayed in a comma-
separated list. While it's not _technically_ ambiguous because
each list item is separated by a comma and a space, a small amount
of extra effort is required to parse "foo.py[2, 1,537, 1,822]".

So, make that "foo.py[2, 1537, 1822]" which is immediately clear.