Fold adjacent Text tokens together in OrgLexer
to avoid emitting a new token for every single word and whitespace
transition.

Also, try to deal a bit more correctly with block state.  I noticed
that the nested block test case originally terminated on the inner
`#+end_...` clause, which is not correct.  My change causes the lexer
to wait until a matching end clause is seen, at the cost of breaking
off the block type as an additional Word token.